improve theming 1
